pyproject.toml
[project]
name = "ollama-text-gateway"
version = "0.1.0"
description = "Simple FastAPI gateway for text generation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0,<1.0.0",
    "uvicorn[standard]>=0.30.0,<1.0.0",
    "httpx>=0.27.0,<1.0.0",
    "sqlalchemy[asyncio]>=2.0.0,<3.0.0",
    "alembic>=1.13.0,<2.0.0",
    "asyncpg>=0.29.0,<1.0.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.0.0,<2.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9.0.0",
    "ruff>=0.6.0,<1.0.0",
    "httpx>=0.27.0,<1.0.0",
]

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["infra*"]
